Moved files around, created new ones, commented code
Hey Jon, I moved the datafiles into their own folder as is general
practice when doing research.  I also split up some of your code, for
instance I created a file for your options and loading the libraries as
well as one to fetch and save your merged dataset.  This makes it easier
for your other scripts as they can just `source()` or `load()` the
options and dataset.

It terms of your guys' code, it is definitely a bit messy! haha its ok
tho. I made some comments throughout the code, affixed with "Jon:"
tags.  So you can do a search for them easily.  In general, you use a
lot of `for` loops and `if..then` statements.  R is not good with for
loops (plus they + the if-then are hard to read).  R's strength is in
being able to vectorize most of its operations (which are much faster!),
plus I would suggest making use of Hadley's dplyr package for the
if-then thingies you have.

Anywho, let me know what you think.